refactor: Extract types out of store
