+++ app/flags/store.ts
@@ -1,20 +1,7 @@
 import mitt, { Emitter } from "mitt";
 
 import lsAdapter from "./ls-adapter";
-
-export type FlagValue = any;
-
-export type FlagName =
-  /** Whether we can search by termsets */
-  | "search.termsets"
-  /** Whether we can add dataset from shared dimensions */
-  | "configurator.add-dataset.shared"
-  /** Whether we can add a new dataset */
-  | "configurator.add-dataset.new"
-  /** Whether we can use the free canvas dashboard layout */
-  | "layouter.dashboard.free-canvas"
-  /** Whether we can use shared filters on dashboard layout */
-  | "layouter.dashboard.shared-filters";
+import { FlagName, FlagValue } from "./types";
 
 type Events = { change: string };
 /**
